refactor styling
