+++ src/components/SkipLink.js
@@ -3,7 +3,7 @@ import { jsx } from "theme-ui";
 
 export default function SkipLink({ href = "#main", children }) {
   return (
-    <a href={href} className="shadow-lg" sx={{ variant: "links.skip" }}>
+    <a href={href} sx={{ variant: "links.skip" }}>
       {children || "Skip to main content"}
     </a>
   );
