+++ src/components/mdx/AnchorHeading.js
@@ -1,5 +1,5 @@
 /** @jsx jsx */
-import { jsx } from "theme-ui";
+import { jsx, Styled } from "theme-ui";
 import * as icons from "components/icons";
 
 // double arrow function iz hard. see: https://stackoverflow.com/a/54839755
@@ -13,16 +13,31 @@ import * as icons from "components/icons";
 // };
 
 export default function getHeading(tag) {
+  function HeadingEl({ tag, ...props }) {
+    switch (tag) {
+      case "h2":
+        return <h2 {...props} />;
+      case "h3":
+        return <h3 {...props} />;
+      case "h4":
+        return <h4 {...props} />;
+      case "h5":
+        return <h5 {...props} />;
+      default:
+        return <h2 {...props} />; // fallback to h2
+    }
+  }
+
   const sxStyle = { variant: `components.mdx.${tag}` };
   return function AnchorHeading({ id, children, ...props }) {
-    if (!id) return <tag sx={sxStyle} {...props} />;
+    if (!id) return <HeadingEl tag={tag} sx={sxStyle} {...props} />;
     return (
-      <tag sx={sxStyle} id={id} {...props}>
+      <HeadingEl tag={tag} sx={sxStyle} id={id} {...props}>
         <a href={`#${id}`}>
           <icons.Link className="invisible absolute" />
           {children}
         </a>
-      </tag>
+      </HeadingEl>
     );
   };
 }
