+++ src/theme/variants/links.js
@@ -1,5 +1,11 @@
 import { basicStyles } from "./buttons";
 
+const backToTop = {
+  mt: 6,
+  display: "inline-block",
+  borderBottom: "2px solid currentColor",
+};
+
 const coverParent = {
   "&::after": {
     content: '""', // https://theme-ui.com/guides/object-styles/#pseudo-elements
@@ -8,6 +14,15 @@ const coverParent = {
     top: 0,
     right: 0,
     bottom: 0,
+  },
+};
+
+const inBody = {
+  color: "primary",
+  borderBottom: "2px solid currentColor",
+  fontWeight: "medium",
+  "&:hover, &:focus": {
+    borderBottomColor: "text",
   },
 };
 
@@ -32,17 +47,12 @@ const skip = {
   },
 };
 
-const backToTop = {
-  mt: 6,
-  display: "inline-block",
-  borderBottom: "2px solid currentColor",
-};
-
 export default {
+  backToTop,
   button: { ...basicStyles }, // Looks like button but is a link.
   coverParent,
+  inBody,
   skip,
-  backToTop,
 };
 
 // a: {
