+++ src/utils/tailwind.css
@@ -1,3 +1,7 @@
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
+
+svg {
+  fill: currentColor;
+}
